Move Tabs utilities to a shared folder
