+++ packages/react-tabs/src/utilities/observeResize.ts
